pyproject.toml
[project]
name = "django-practice"
version = "1.0.0"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "celery==5.4.0",
    "django==5.0.6",
    "django-bulk-sync==3.3.0",
    "django-celery-beat==2.6.0",
    "django-cors-headers==4.3.1",
    "django-extensions==3.2.3",
    "django-filter==24.2",
    "django-guid==3.5.0",
    "django-imagekit==5.0.0",
    "django-simple-history==3.8.0",
    "django-storages==1.14.3",
    "djangorestframework==3.15.1",
    "djangorestframework-camel-case==1.4.2",
    "djangorestframework-simplejwt==5.4.0",
    "drf-extensions==0.7.1",
    "drf-nested-routers==0.94.1",
    "drf-spectacular==0.27.2",
    "gunicorn==23.0.0",
    "myst-parser==4.0.1",
    "psycopg2-binary==2.9.10",
    "pyhumps==3.8.0",
    "pyjwt==2.8.0",
    "python-dotenv==1.0.1",
    "redis==5.1.1",
    "requests==2.32.3",
    "sphinx==8.1.3",
    "sphinx-rtd-theme==3.0.2",
    "sphinxcontrib-mermaid==1.0.0",
    "watchtower==3.2.0",
]

[dependency-groups]
dev = [
    "ruff==0.11.2",
    "pre-commit==3.7.1",
    "pytest-xdist==3.6.1",
    "pytest-django==4.8.0",
    "factory-boy==3.3.0",
    "testcontainers[postgres]==4.9.1",
    "testcontainers[redis]==4.9.1",
]

[tool.pytest.ini_options]
django_debug_mode = true
addopts = "--ds=settings"
testpaths = ["tests"]

[tool.ruff]
line-length = 120

exclude = [
    "migrations",
]

[tool.ruff.lint]
ignore = [
    "F403",  # not from module import *
    "F405",  # defined from star imports
    "E501",  # Line too long
    "E722",  # Do not use bare except, specify exception instead
    "B904",  # Within an `except` clause, raise exceptions
    "SIM108",  # if else -> one line
    "SIM102"  # Use a single `if` statement instead of nested `if` statements
]

extend-select = [
    "TC",
    "E",  # pycodestyle
    "F",  # pyflakes
    "I",  # isort
    "UP", # pyupgrade
    "B",  # bugbear
    "C4", # comprehensions
    "SIM", # simplify
]


[tool.ruff.lint.isort]
combine-as-imports = true
lines-after-imports = 2


[tool.ruff.format]
docstring-code-format = true
